+++ app/src/main/java/com/example/smartalarm/adapter/AlarmListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.smartalarm;
+package com.example.smartalarm.adapter;
 
 import android.app.Activity;
 import android.content.Context;
@@ -10,6 +10,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.RecyclerView;
+import com.example.smartalarm.R;
 import com.example.smartalarm.dataStructures.NameIdPair;
 
 import java.util.HashMap;
